+++ bot/bot.py
@@ -11,7 +11,7 @@ import datetime
 import os
 
 import bot_funcs as funcs
-import utility.enums as enums
+import utility.configs as configs
 from utility.setup_logger import logger 
 
 
@@ -71,7 +71,7 @@ async def heartbeat_handler(ctx):
         help="Enlist a member and grant basic recruit roles. Accepts a single member mention as an argument. Will fail when attempting to enlist a member with the 87th Regiment of Foot role.\nUsage: !enlist @<member>",
         brief="Enlist a member and grant basic recruit roles"
         )
-@has_role(enums.BOT_USER_ROLE)
+@has_role(configs.BOT_USER_ROLE)
 async def enlist_member_handler(ctx, user: discord.Member):
     await funcs.enlist_member(ctx, user)
 
@@ -91,7 +91,7 @@ async def enlist_error(ctx, error):
         help="Grant Merc/Rep/Visitor roles. Accepts a single member mention as an argument. Role type ('rep', 'merc', or 'visitor') must be defined or else command will fail. \nUsage: !grantrole merc/rep/visitor @<member>",
         brief="Adds Merc, Rep, or Visitor roles to a member"
         )
-@has_role(enums.BOT_USER_ROLE)
+@has_role(configs.BOT_USER_ROLE)
 async def grant_role_handler(ctx, roleType, user: discord.User):
     await funcs.grant_role(ctx, roleType, user)
 
@@ -111,7 +111,7 @@ async def grant_role_error(ctx, error):
         help="Update the master document's 'Last Seen' column with members currently in the voice channel. Can be used in any text channel the bot has access to, but the invoker must be present in the voice call.",
         brief="Take an attendance count"
         )
-@has_role(enums.BOT_USER_ROLE)
+@has_role(configs.BOT_USER_ROLE)
 async def attend_handler(ctx):
     await funcs.attend(ctx)