+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Note: This file was started on 2023-04-22. Any changes before then are not captu
 
 ## 2023-12-10
 - Updated README
+- Renamed `enums.py` to `configs.py` to better represent content of file
 
 ## 2023-12-09
 - Changed name of signup function to better reflect its purpose (`check_and_put_signup_reactions` -> `add_or_remove_signup_reactions`)